Prefer "authored" commits over "reviewed" commits.
The case of an author inadvertently hitting CR+1 is likely to be
much more prevalent than the case of a reviewer taking action on the CL
that results in them inadvertently becoming an author.

This patches ignore reviewer that are also the author of the patch.

This ensures the following invariant:
The sum of everyone's authored patchs equal the sum of patches.

Bug: https://github.com/ArthurSonzogni/ChromeCommitTracker/issues/7