Fetch org-list.txt from chromium.org.
Reported by @wfh@chromium.org. Thanks!
The `org-list.txt` file is meant to be fetched from the Chromium
repository. However, this line was commented in the code to allow for
the "Ladybird" organization to be added to the list of organizations
locally without affecting the Chromium repository.

I forgot to uncomment this line and "undo" the local modification.